requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Statistical analysis
statsmodels>=0.13.0

# Utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Development
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0
flake8>=3.9.0
